scripts for making a flux diagram with pictures
run overall_flux_diagram.py with "use_smiles" as false. this makes a dot file with the species names
then, run prettydot.py. if this gives you an error, the dot build you have may be missing a package. depending on
the formats sometimes I have to get out of the conda environment and run "dot" for it to work.